+++ utils/test-utils/src/main/java/de/sovity/edc/extension/e2e/extension/E2eScenario.java
@@ -113,6 +113,10 @@ public class E2eScenario {
         return internalCreateAsset(id, uiDataSource).getId();
     }
 
+    public String createAsset(UiAssetCreateRequest uiAssetCreateRequest) {
+        return providerClient.uiApi().createAsset(uiAssetCreateRequest).getId();
+    }
+
     public MockedAsset createAssetWithMockResource(String id) {
 
         val path = "/assets/" + id;